feat: add test for GET s3URL, passing. Also moved s3 to own files, controlle and route
