+++ package.json
@@ -37,7 +37,6 @@
     "@types/mongoose": "^5.11.97",
     "aws-sdk": "^2.1228.0",
     "cors": "^2.8.5",
-    "crypto": "^1.0.1",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "mongodb": "^4.8.1",
